requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
ultralytics
torch
opencv-python
gunicorn
python-dotenv
watchdog